javaCC working

+++ src/main/java/org/example/Actions/FileManager.java
@@ -12,8 +12,13 @@ import java.nio.file.Path;
 public class FileManager {
     private Path filePath;
     private boolean isFileSaved = true;
-    public File openFileChooser(Component parent, Interface classParent) {
-        if(saveInSecondChance(classParent)){
+    private Interface parent;
+    private boolean isFileOpened = false;
+    public FileManager(Interface parent) {
+        this.parent = parent;
+    }
+    public File openFileChooser() {
+        if(saveInSecondChance()){
             JFileChooser fileChooser = new JFileChooser(new java.io.File("resources"));
             fileChooser.setPreferredSize(new Dimension(800, 600));
             fileChooser.setFileSelectionMode(JFileChooser.FILES_ONLY);
@@ -26,7 +31,8 @@ public class FileManager {
                 if (selectedFile.getPath().endsWith(".txt")) {
                     this.filePath = selectedFile.toPath();
                     this.isFileSaved = true;
-                    classParent.setWindowTitle("Compilador - "+selectedFile.getName());
+                    parent.setWindowTitle("Compilador - "+selectedFile.getName());
+                    isFileOpened = true;
                     return selectedFile;
                 } else {
                     JOptionPane.showMessageDialog(parent,
@@ -40,35 +46,47 @@ public class FileManager {
         return null;
     }
 
-    public int saveFile(Interface parent) {
-        JFileChooser fileChooser = new JFileChooser(new java.io.File("resources"));
-        fileChooser.setPreferredSize(new Dimension(800, 600));
-        fileChooser.setFileSelectionMode(JFileChooser.DIRECTORIES_ONLY);
-        fileChooser.setDialogTitle("Selecione o diretorio para salvar");
-        fileChooser.setApproveButtonText("Salvar");
+    public int saveFile() {
+        if(!isFileOpened){
+            JFileChooser fileChooser = new JFileChooser(new java.io.File("resources"));
+            fileChooser.setPreferredSize(new Dimension(800, 600));
+            fileChooser.setFileSelectionMode(JFileChooser.DIRECTORIES_ONLY);
+            fileChooser.setDialogTitle("Selecione o diretorio para salvar");
+            fileChooser.setApproveButtonText("Salvar");
 
-        int returnValue = fileChooser.showOpenDialog(parent);
-        if (returnValue == JFileChooser.APPROVE_OPTION) {
-            File selectedDir = fileChooser.getSelectedFile();
-            if(!selectedDir.getName().toLowerCase().endsWith(".txt")){
-                selectedDir = new File(selectedDir.getAbsolutePath()+".txt");
+            int returnValue = fileChooser.showOpenDialog(parent);
+            if (returnValue == JFileChooser.APPROVE_OPTION) {
+                File selectedDir = fileChooser.getSelectedFile();
+                if(!selectedDir.getName().toLowerCase().endsWith(".txt")){
+                    selectedDir = new File(selectedDir.getAbsolutePath()+".txt");
+                }
+                try(FileWriter fw = new FileWriter(selectedDir)) {
+                    fw.write(parent.getTextInput().getText());
+                    parent.setWindowTitle("Compilador"+" - "+selectedDir.getName());
+                    JOptionPane.showMessageDialog(parent,"Arquivo salvo em "+selectedDir.getAbsolutePath()+" com sucesso!");
+                    this.isFileSaved = true;
+                    this.filePath = selectedDir.toPath();
+                }catch (Exception e){
+                    e.printStackTrace();
+                }
+                isFileOpened = true;
+                return 1;
             }
-            try(FileWriter fw = new FileWriter(selectedDir)) {
+        }else{
+            try(FileWriter fw = new FileWriter(filePath.toFile())) {
                 fw.write(parent.getTextInput().getText());
-                parent.setWindowTitle("Compilador"+" - "+selectedDir.getName());
-                JOptionPane.showMessageDialog(parent,"Arquivo salvo em "+selectedDir.getAbsolutePath()+" com sucesso!");
+                parent.setWindowTitle("Compilador"+" - "+filePath.toFile().getName());
+                JOptionPane.showMessageDialog(parent,"Arquivo salvo em "+filePath.toFile().getAbsolutePath()+" com sucesso!");
                 this.isFileSaved = true;
-                this.filePath = selectedDir.toPath();
+                this.filePath = filePath.toFile().toPath();
             }catch (Exception e){
                 e.printStackTrace();
             }
-
-            return 1;
         }
         return 0;
     }
 
-    public boolean saveInSecondChance(Interface parent) {
+    public boolean saveInSecondChance() {
         if (!this.isFileSaved) {
             int option = JOptionPane.showConfirmDialog(
                     parent,
@@ -78,9 +96,9 @@ public class FileManager {
                     JOptionPane.QUESTION_MESSAGE
             );
             if (option == JOptionPane.YES_OPTION){
-                int file = saveFile(parent);
+                int file = saveFile();
                 if( file != 1){
-                    saveInSecondChance(parent);
+                    saveInSecondChance();
                 }
                 return true;
             }else{
@@ -102,7 +120,7 @@ public class FileManager {
         return isFileSaved;
     }
 
-    public void setFileSaved(boolean fileSaved, Interface parent) {
+    public void setFileSaved(boolean fileSaved) {
         isFileSaved = fileSaved;
         if (parent.getWindowTitle().endsWith("*") && isFileSaved) {
             parent.setWindowTitle(parent.getWindowTitle().replace("*", ""));
